Controller Examen ?

+++ Badge2022EF.WebApi/Controllers/ExamensController.cs
@@ -1,5 +1,7 @@
 ﻿using Badge2022EF.DAL.Repositories;
+using Badge2022EF.DAL.Repositories.Interface;
 using Badge2022EF.Models.Concretes;
+using Badge2022EF.Models.Interfaces;
 using Badge2022EF.WebApi.Filters;
 using Badge2022EF.WebApi.Models;
 using Microsoft.AspNetCore.Mvc;
@@ -37,48 +39,51 @@ namespace Badge2022EF.WebApi.Controllers
         // GET api/<ExamensController>/5
         [HttpGet("{id}")]
         [Authorization("Admin", "Praticien", "Patient")]
-        public Examens GetOne(int id)
+        public IEnumerable<Examens> GetOne(int id)
         {
-            return _ExamenRepository.GetOne(id);
+            IEnumerable<Examens> aa = _ExamenRepository.GetOne2(id);
+            foreach (var item in aa) { };
+            return aa.AsEnumerable().Where(x => x.eid == id);
         }
 
         // POST api/<ExamensController>
         [HttpPost]
         [Authorization("Admin", "Praticien")]
-        public void Post([FromBody] J_Examens newExamen)
+        public async Task<IActionResult> Post([FromBody] J_Examens newExamen)
         {
-            Examens Examen = new(
-                        newExamen.eid,
-                        newExamen.enom,
-                        newExamen.enote
-                        )
-            {
-                eid = 0
-            };
+            Examens Examen = new(newExamen.eid, newExamen.enom, newExamen.enote);
             _ExamenRepository.Add(Examen);
+            return Ok();
         }
 
         // PUT api/<ExamensController>/5
         [HttpPut("{id}")]
         [Authorization("Admin", "Praticien")]
-        public void Put(long id, [FromBody] J_Examens majExamen)
+        public async Task<IActionResult> Put(int id, [FromBody] J_Examens majExamen)
         {
-            Examens Examen = new(
-                        majExamen.eid,
-                        majExamen.enom,
-                        majExamen.enote
-                        )
+            Examens ar = _ExamenRepository.GetOne(id);
+            if (ar.eid == id)
             {
-            };
-            _ExamenRepository.Update(Examen);
+                Examens Examen = new(majExamen.eid, majExamen.enom, majExamen.enote);
+                _ExamenRepository.Update(Examen);
+                return Ok();
+            }
+            return BadRequest();
+
         }
 
         // DELETE api/<ExamensController>/5
         [HttpDelete("{id}")]
         [Authorization("Admin", "Praticien")]
-        public void Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            _ExamenRepository.Delete(id);
+            Examens ar = _ExamenRepository.GetOne(id);
+            if (ar.eid == id)
+            {
+                _ExamenRepository.Delete(id);
+                return Ok();
+            }
+            return BadRequest();
         }
     }
 }
